+++ src/components/AdminQuestions/AdminQuestions.test.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import AdminSearchStudentContainer from './AdminSearchStudentContainer';
+import AdminQuestions from './AdminQuestions';
 import { MockedProvider } from '@apollo/react-testing';
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 configure({ adapter: new Adapter() });
 
-describe('AdminSearchStudentContainer', () => {
+describe('AdminAllQuestionsContainer', () => {
   let wrapper;
 
   beforeEach(() => {
     wrapper = render(
       <MockedProvider addTypeName={false}>
-        <AdminSearchStudentContainer />
+        <AdminQuestions/>
       </MockedProvider>
     );
   });
