--- src/components/SelectStudentContainer/SelectStudentContainer.test.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import SelectStudentContainer from './SelectStudentContainer';
-import { MockedProvider } from '@apollo/react-testing';
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-configure({ adapter: new Adapter() });
-
-describe('SelectStudentContainer', () => {
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = render(
-      <MockedProvider addTypeName={false}>
-        <SelectStudentContainer />
-      </MockedProvider>
-    );
-  });
-
-  it('should match the snapshot with all data passed in correctly', () => {
-    expect(wrapper).toMatchSnapshot();
-  });
-});
